requirements.txt
# Core data processing
pandas>=2.2.0
openpyxl>=3.1.2
xlrd>=2.0.1
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.4.0
torch>=2.2.0
transformers>=4.38.0
sentence-transformers>=3.0.0

# Vector database - using simpler alternative for Python 3.14 compatibility
faiss-cpu>=1.8.0

# Visualization
plotly>=5.18.0
umap-learn>=0.5.5

# Web interface
streamlit>=1.32.0
